requirements.txt
tqdm
imageio
pillow
pandas
numpy
opencv-python
picamera2
matplotlib
clr